🔨 Refactor: 펀딩 개설 페이지 컴포넌트 분리
