+++ src/constants/constants.ts
@@ -1,0 +1,1 @@
+export const DRAWER_BLEEDING = 56;
